pyproject.toml
[project]
name = "computational-thinking"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "folium>=0.20.0",
    "geopy>=2.4.1",
    "googlemaps>=4.10.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pydeck>=0.9.1",
    "python-dotenv>=1.2.1",
    "seaborn>=0.13.2",
    "streamlit>=1.51.0",
    "streamlit-folium>=0.25.3",
]
